first pass at tokio metrics
