+++ runtime/src/deterministic.rs
@@ -25,6 +25,7 @@
 //! println!("Auditor state: {}", auditor.state());
 //! ```
 
+use crate::metrics::Metrics;
 use crate::{Clock, Error, Handle};
 use bytes::Bytes;
 use futures::{
@@ -33,11 +34,7 @@ use futures::{
     SinkExt, StreamExt,
 };
 use governor::clock::{Clock as GClock, ReasonablyRealtime};
-use prometheus_client::{
-    encoding::EncodeLabelSet,
-    metrics::{counter::Counter, family::Family},
-    registry::Registry,
-};
+use prometheus_client::registry::Registry;
 use rand::{prelude::SliceRandom, rngs::StdRng, CryptoRng, RngCore, SeedableRng};
 use sha2::{Digest, Sha256};
 use std::{
@@ -52,37 +49,6 @@ use std::{
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
 use tracing::trace;
-
-#[derive(Clone, Debug, Hash, PartialEq, Eq, EncodeLabelSet)]
-struct Link {
-    origin: String,
-    destination: String,
-}
-
-#[derive(Debug)]
-struct Metrics {
-    tasks_spawned: Counter,
-    task_polls: Counter,
-    bandwidth: Family<Link, Counter>,
-}
-
-impl Metrics {
-    fn record_task_spawned(&self) {
-        self.tasks_spawned.inc();
-    }
-
-    fn record_task_poll(&self) {
-        self.task_polls.inc();
-    }
-
-    fn record_bandwidth(&self, origin: SocketAddr, destination: SocketAddr, bytes: usize) {
-        let link = Link {
-            origin: origin.to_string(),
-            destination: destination.to_string(),
-        };
-        self.bandwidth.get_or_create(&link).inc_by(bytes as u64);
-    }
-}
 
 /// Range of ephemeral ports assigned to dialers.
 const EPHEMERAL_PORT_RANGE: Range<u16> = 32768..61000;
@@ -274,11 +240,7 @@ pub struct Executor {
 impl Executor {
     /// Initialize a new `deterministic` runtime with the given seed and cycle duration.
     pub fn init(cfg: Config) -> (Runner, Context, Arc<Auditor>) {
-        let metrics = Arc::new(Metrics {
-            tasks_spawned: Counter::default(),
-            task_polls: Counter::default(),
-            bandwidth: Family::default(),
-        });
+        let metrics = Arc::new(Metrics::init(cfg.registry));
         let auditor = Arc::new(Auditor::new());
         let executor = Arc::new(Self {
             cycle: cfg.cycle,
@@ -292,24 +254,6 @@ impl Executor {
             }),
             sleeping: Mutex::new(BinaryHeap::new()),
         });
-        {
-            let mut registry = cfg.registry.lock().unwrap();
-            registry.register(
-                "tasks_spawned",
-                "Total number of tasks spawned",
-                metrics.tasks_spawned.clone(),
-            );
-            registry.register(
-                "task_polls",
-                "Total number of task polls",
-                metrics.task_polls.clone(),
-            );
-            registry.register(
-                "bandwidth",
-                "Bandwidth usage by origin and destination",
-                metrics.bandwidth.clone(),
-            );
-        }
         (
             Runner {
                 executor: executor.clone(),
@@ -791,60 +735,6 @@ mod tests {
     use crate::utils::run_tasks;
     use futures::task::noop_waker;
     use std::net::{IpAddr, Ipv4Addr};
-
-    #[test]
-    fn test_bandwidth_metrics() {
-        let cfg = Config::default();
-        let (runner, _, _) = Executor::init(cfg);
-        let metrics = runner.executor.metrics.clone();
-
-        // Send some data
-        let socket1 = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 8080);
-        let socket2 = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 8081);
-        metrics.record_bandwidth(socket1, socket2, 100);
-        metrics.record_bandwidth(socket2, socket1, 200);
-        metrics.record_bandwidth(socket2, socket1, 150);
-
-        // Verify tracking
-        assert_eq!(
-            metrics
-                .bandwidth
-                .get_or_create(&Link {
-                    origin: socket1.to_string(),
-                    destination: socket2.to_string(),
-                })
-                .get(),
-            100
-        );
-        assert_eq!(
-            metrics
-                .bandwidth
-                .get_or_create(&Link {
-                    origin: socket2.to_string(),
-                    destination: socket1.to_string(),
-                })
-                .get(),
-            350
-        );
-    }
-
-    #[test]
-    fn test_task_metrics() {
-        let cfg = Config::default();
-        let (runner, _, _) = Executor::init(cfg);
-        let metrics = runner.executor.metrics.clone();
-
-        for _ in 0..5 {
-            metrics.record_task_spawned();
-        }
-
-        for _ in 0..10 {
-            metrics.record_task_poll();
-        }
-
-        assert_eq!(metrics.tasks_spawned.get(), 5);
-        assert_eq!(metrics.task_polls.get(), 10);
-    }
 
     fn run_with_seed(seed: u64) -> (String, Vec<usize>) {
         let cfg = Config {
